Other character art,  cleaner conversation code, basic move around and talk to peopel
like version 0.3ish cuz no multiple dialogues or game flags and no ending, but you cna move around and talk to people, and the core idea of the mystery is there. Would just need to refactor the dialogue structure a little bit and build a lot more conversations to fill the rest of the story. Main functionality missing that I would like would be a decision/dialogue option set of boxes so we can choose what to bring up. doesn't make a huge amount of sense otherwise, but somewhat close to a mvp of a 3 minute game lol